Switched from TreeView to ListBox for listing the plugins in the MainWindow.
First reason is that the list provided there is labeled as as "Plugins" so it should only contain plugins.
Second reason is that since it's possible to select the commands listed under a plugin in the treeview, it breaks the binding to the PluginSettingsEditor-control.

If it will appear required in the future to see the commands provided by a plugin, I will have to find a different way to do so than by using a TreeView.